backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
matplotlib
seaborn
google-generativeai
psycopg2-binary
sqlalchemy
python-dotenv
mcp
httpx
